raspi_py_uart_send/pyproject.toml
[project]
name = "raspi-py-uart-send"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyserial>=3.5",
]
